Feat: 로그인 중이면, 로그인 페이지 이동 불가 알림 설정
