+++ FE/src/pages/LoginPage.js
@@ -6,6 +6,7 @@ import ButtonSmall from '../components/ButtonSmall.js';
 import { formDataToJson } from '../utils/formDataToJson.js';
 import apiEndpoints from '../constants/apiConfig.js';
 import { CLIENT_ID, REDIRECT_URI } from '../constants/constants.js';
+import { browserInfo } from '../utils/browserInfo.js';
 
 const html = String.raw;
 
@@ -74,6 +75,7 @@ class LoginPage {
 			.classList.add('modal-hidden');
 		fetch(apiEndpoints.LOGIN_CHECK_URL, { method: 'GET' }).then((res) => {
 			if (res.status === 200) {
+				browserInfo('로그인 페이지로 이동이 불가합니다.');
 				changeUrl('playMode');
 			}
 		});
